Add moves_available to Piece and organize code
