+++ spec/game_status_checker_spec.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+RSpec.describe GameStatusChecker do
+end
